+++ lib/getIntentByQuery.js
